Fix rubocop issue, long module, by extracting StripeStubs to a separate helper module
