+++ spec/controllers/spree/admin/orders/payments/payments_controller_refunds_spec.rb
@@ -5,6 +5,7 @@ require 'spree/core/gateway_error'
 
 describe Spree::Admin::PaymentsController, type: :controller do
   include StripeHelper
+  include StripeStubs
 
   let!(:shop) { create(:enterprise) }
   let!(:user) { shop.owner }
